Refactored Socket class to become an interface and existing implementation now became TcpSocket. Also, updated the test server version to 3.9-SNAPSHOT.
