+++ hazelcast/src/hazelcast/client/connection/InSelector.cpp
@@ -18,13 +18,13 @@
 //
 
 #include <string.h>
+#include <errno.h>
 
 #include "hazelcast/util/ILogger.h"
 #include "hazelcast/client/connection/InSelector.h"
 #include "hazelcast/client/connection/ReadHandler.h"
 #include "hazelcast/client/connection/ConnectionManager.h"
 #include "hazelcast/client/connection/Connection.h"
-#include "hazelcast/util/Thread.h"
 
 #if  defined(WIN32) || defined(_WIN32) || defined(WIN64) || defined(_WIN64)
 #pragma warning(push)
