+++ hazelcast/src/hazelcast/client/connection/OutSelector.cpp
@@ -18,11 +18,12 @@
 //
 
 #include <string.h>
+#include <errno.h>
+
 #include "hazelcast/client/connection/OutSelector.h"
 #include "hazelcast/client/connection/ConnectionManager.h"
 #include "hazelcast/client/connection/Connection.h"
 #include "hazelcast/util/ILogger.h"
-#include "hazelcast/util/Thread.h"
 
 #if  defined(WIN32) || defined(_WIN32) || defined(WIN64) || defined(_WIN64)
 #pragma warning(push)
